test_requirements.txt
django-nose
django-libs
ipdb
flake8
fabric
factory_boy
mock
